스터디 관리 기능 구현 완료, CSS 작업 중 (#287)
* 토큰 요청 리팩토링

* 리팩토링 수정 완료

* 병합 전 커밋

* 마이페이지 구현 및 테스트 완료

* 마이페이지 구현 완료 및 리스트 오류 해결 중

* 멤버 목록 조회

* 가입된 멤버조회 구현

* 마이페이지 / 스터디 리스트 페이지 테스트 완료

* 개인 일정만 구현하면 진짜진짜 끝...

* 세부사양 구현 완료, 개인 일정 CRUD의 경우 엔드포인트가 확정되면 추후 개선

* 기능구현 완료, 디버깅 및 CSS 구현 중

* 기능구현 완료 및 CSS 구현 중

* 세부사양 구현 완료 및 CSS 수정 중

* fix : 리더 권한 위임

* 스터디장 탈퇴 block 처리

* 비밀번호 검증 로직 수정중

* 기능구현 최종, CSS 작업중

+++ client/src/pages/ProfileInfo.tsx
@@ -6,14 +6,14 @@ import {
   updateMemberDetail,
   MemberDetailDto,
   deleteMember,
-  MemberPasswordCheckDto,
-  checkMemberPassword,
+  checkOauth2Member,
 } from "../apis/MemberApi";
 import { useState, useEffect, ChangeEvent } from "react";
 import UserInfoEditModal from "../components/modal/UserInfoEditModal";
 import { useRecoilValue } from "recoil";
 import { LogInState } from "../recoil/atoms/LogInState";
 import { useNavigate } from "react-router-dom";
+import CheckPasswordModal from "../components/modal/CheckPasswordModal";
 
 const ProfileInfo = () => {
   const isLoggedIn = useRecoilValue(LogInState);
@@ -27,6 +27,8 @@ const ProfileInfo = () => {
   });
   // 멤버 정보 수정 (클라이언트에서 수정된 데이터)
   const [isIntroduceEdit, setIsIntroduceEdit] = useState<boolean>(false);
+  const [passowrdCheckModalOpen, setPasswordCheckModalOpen] =
+    useState<boolean>(false);
   const navigate = useNavigate();
 
   // TODO 최초 페이지 진입 시 유저의 정보를 조회하는 코드
@@ -45,19 +47,13 @@ const ProfileInfo = () => {
 
   // TODO Edit 버튼을 클릭 시, 유저의 닉네임, 비밀번호를 수정할 수 있도록 상태를 변경하는 코드
   // 현재 Modal 구현은 완료했으나 비동기 처리로 인해 계속된 오류 발생. 추가적인 최적화 작업 요함
+  // Jest로 테스트할 필요! : why? 소셜 로그인은 자동으로 배포 서버로 리다이렉션 함!
   const handleEditClick = async () => {
-    const enteredPassword = prompt(
-      "개인정보 수정 전 비밀번호를 확인해야 합니다."
-    );
-    if (!enteredPassword) return; // 비밀번호 입력을 취소하면 함수 종료
-    try {
-      const passwordCheckDto: MemberPasswordCheckDto = {
-        password: enteredPassword,
-      };
-      await checkMemberPassword(passwordCheckDto);
-      setIsModalOpen(true); // 비밀번호 검증이 성공하면 모달 열기
-    } catch (error) {
-      alert("비밀번호가 일치하지 않습니다.");
+    const data = await checkOauth2Member(isLoggedIn);
+    if (data.provider !== "LOCAL") {
+      alert("소셜 로그인 유저는 개인정보를 수정할 수 없습니다.");
+    } else {
+      setPasswordCheckModalOpen(true);
     }
   };
 
@@ -93,15 +89,19 @@ const ProfileInfo = () => {
   // TODO DELETE 버튼을 클릭 시, 유저의 자기소개 및 원하는 동료상을 서버에서 DELETE하는 코드
   const handleDeleteClick = async () => {
     try {
-      await deleteMember();
-      alert("회원탈퇴가 완료되었습니다.");
-      localStorage.clear();
-      navigate("/");
+      const confirmed = window.confirm("정말로 회원탈퇴하시겠습니까?");
+      if (confirmed) {
+        await deleteMember();
+        alert("회원탈퇴가 완료되었습니다.");
+        localStorage.clear();
+        navigate("/");
+      } else {
+        alert("회원탈퇴가 취소되었습니다.");
+      }
     } catch (error) {
       console.error(error);
     }
   };
-
   return (
     <Wrapper>
       <ProfileBaseWrapper>
@@ -115,13 +115,11 @@ const ProfileInfo = () => {
           <EditButton onClick={handleEditClick}>Edit</EditButton>
         </ProfileBaseInfo>
       </ProfileBaseWrapper>
-
-      {/* 유저의 자기소개와 원하는 유형의 팀원을 정리하는 자리 */}
       <IntroduceAndDesired>
         {!isIntroduceEdit ? (
           <>
-            <IntroduceAndDesiredInput value={memberInfo?.aboutMe} />
-            <IntroduceAndDesiredInput value={memberInfo?.withMe} />
+            <IntroduceAndDesiredInput value={memberInfo?.aboutMe} disabled />
+            <IntroduceAndDesiredInput value={memberInfo?.withMe} disabled />
           </>
         ) : (
           <>
@@ -137,16 +135,23 @@ const ProfileInfo = () => {
             />
           </>
         )}
-        {!isIntroduceEdit ? (
-          <EditButton onClick={handleIntroduceEditClick}>Edit</EditButton>
-        ) : (
-          <EditButton onClick={handleSaveClick}>Save</EditButton>
-        )}
-        <EditButton onClick={handleDeleteClick}>Delete</EditButton>
+        <ButtonWrapper>
+          <ExitEditButton onClick={handleDeleteClick}>회원탈퇴</ExitEditButton>
+          {!isIntroduceEdit ? (
+            <EditButton onClick={handleIntroduceEditClick}>Edit</EditButton>
+          ) : (
+            <EditButton onClick={handleSaveClick}>Save</EditButton>
+          )}
+        </ButtonWrapper>
       </IntroduceAndDesired>
       <UserInfoEditModal
         isOpen={isModalOpen}
         closeModal={() => setIsModalOpen(false)}
+      />
+      <CheckPasswordModal
+        isOpen={passowrdCheckModalOpen}
+        closeModal={() => setPasswordCheckModalOpen(false)}
+        setIsModalOpen={setIsModalOpen}
       />
     </Wrapper>
   );
@@ -154,11 +159,72 @@ const ProfileInfo = () => {
 
 export default ProfileInfo;
 
-const Wrapper = styled.div``;
-const ProfileBaseWrapper = styled.div``;
-const ProfileImage = styled.div``;
-const ProfileBaseInfo = styled.div``;
-const IntroduceAndDesired = styled.div``;
-const ProfileInput = styled.input``;
-const IntroduceAndDesiredInput = styled.input``;
-const EditButton = styled.button``;
+const Wrapper = styled.div`
+  width: 100%;
+  max-width: 960px;
+  margin: 0 auto;
+`;
+
+const ProfileBaseWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 20px;
+`;
+
+const ProfileImage = styled.div`
+  margin-right: 20px;
+`;
+
+const ProfileInput = styled.input`
+  margin-bottom: 10px;
+  padding: 8px;
+  width: 100%;
+`;
+
+const ProfileBaseInfo = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+`;
+
+const IntroduceAndDesired = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  margin-top: 20px;
+  width: 100%;
+`;
+
+const IntroduceAndDesiredInput = styled.input`
+  margin-bottom: 10px;
+  padding: 8px;
+  width: 100%;
+  height: 200px;
+`;
+
+const ButtonWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+`;
+
+const EditButton = styled.button`
+  margin-bottom: 10px;
+  padding: 8px 16px;
+  background-color: #4d74b1;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+`;
+
+const ExitEditButton = styled.button`
+  margin-bottom: 10px;
+  padding: 8px 16px;
+  background-color: #7e1717;
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+`;
